# modulo_promocional_api/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Modulos-Promocionales-api
  description: Microservicio de gestion de peticiones de Informacion e inyeccion de
    Datos
  version: 1.0.11
servers:
- url: /rest/modulos-promocionales-api/v1.0/
...
paths:
  /get/lugares/provincias:
    post:
      tags:
      - Lugares
      description: Consulta de Datos de Provincias segun sus parametros.
      operationId: get_provincias
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetProvincia'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetProvincia_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.lugares_controller
  /get/lugares/ciudades:
    post:
      tags:
      - Lugares
      description: Consulta de Datos de Ciudades
      operationId: get_ciudades
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetCiudades'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetCiudad_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.lugares_controller
  /get/lugares/sectores:
    post:
      tags:
      - Lugares
      description: Consulta de Datos de Sectores
      operationId: get_sectores
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetSectores'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetSectores_Data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.lugares_controller
...
components:
  schemas:
    RequestGetProvincia:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_PROV
        TARIFFPLANVARIANT:
          type: integer
          example: 200
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetCiudades:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_CITY
        id_Prov:
          type: integer
          example: 12123
        id_Provs:
          type: array
          example:
          - 12123
          - 45678
          - 78901
          items:
            type: integer
        TARIFFPLANVARIANT:
          type: integer
          example: 200
        PRODUCTOID:
          type: integer
          example: 200
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetSectores:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_SECTORS
        id_City:
          type: integer
          example: 12123
        id_Cities:
          type: array
          example:
          - 12123
          - 45678
          - 78901
          items:
            type: integer
        TARIFFPLANVARIANT:
          type: integer
          example: 200
        PRODUCTOID:
          type: integer
          example: 200
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
...
    ResponseGetProvincia_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        PROVINCIES:
          $ref: '#/components/schemas/ResponseGetProvincia_data_PROVINCIES'
      example:
        PROVINCIES:
          PROVINCIA: PROVINCIA
          PROVINCIA_ID: 0
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
    ResponseGetCiudad_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        CITIES:
          $ref: '#/components/schemas/ResponseGetCiudad_data_CITIES'
        CITIESxPROV:
          $ref: '#/components/schemas/ResponseGetCiudad_data_CITIES'
      example:
        errorCode: 0
        CITIES:
          PROVINCIA: PROVINCIA
          CIUDAD: CIUDAD
          CIUDAD_ID: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
    ResponseGetSectores_Data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        SECTORS:
          $ref: '#/components/schemas/ResponseGetSectores_Data_SECTORS'
        SECTORSxCITY:
          $ref: '#/components/schemas/ResponseGetSectores_Data_SECTORS'
      example:
        SECTORS:
          SECTOR: SECTOR
          CIUDAD: CIUDAD
          SECTOR_ID: 0
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
...
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
          example: -1
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: mensaje de error
...
    ResponseGetProvincia_data_PROVINCIES:
      type: object
      properties:
        PROVINCIA_ID:
          type: integer
        PROVINCIA:
          type: string
      example:
        PROVINCIA: PROVINCIA
        PROVINCIA_ID: 0
    ResponseGetCiudad_data_CITIES:
      type: object
      properties:
        CIUDAD_ID:
          type: integer
        CIUDAD:
          type: string
        PROVINCIA:
          type: string
      example:
        PROVINCIA: PROVINCIA
        CIUDAD: CIUDAD
        CIUDAD_ID: 0
    ResponseGetSectores_Data_SECTORS:
      type: object
      properties:
        SECTOR_ID:
          type: integer
        SECTOR:
          type: string
        CIUDAD:
          type: string
      example:
        SECTOR: SECTOR
        CIUDAD: CIUDAD
        SECTOR_ID: 0
...